Refactored generator template helper
