+++ tests/unit/TransferGenerator/Generator/Render/TemplateRenderTest.php
@@ -10,8 +10,9 @@ use PHPUnit\Framework\TestCase;
 use Picamator\TransferObject\Generated\DefinitionContentTransfer;
 use Picamator\TransferObject\Generated\TemplateTransfer;
 use Picamator\TransferObject\TransferGenerator\Exception\TransferGeneratorException;
-use Picamator\TransferObject\TransferGenerator\Generator\Render\Template\TemplateHelper;
+use Picamator\TransferObject\TransferGenerator\Generator\Enum\TransferEnum;
 use Picamator\TransferObject\TransferGenerator\Generator\Render\TemplateBuilderInterface;
+use Picamator\TransferObject\TransferGenerator\Generator\Render\TemplateHelper;
 use Picamator\TransferObject\TransferGenerator\Generator\Render\TemplateRender;
 use Picamator\TransferObject\TransferGenerator\Generator\Render\TemplateRenderInterface;
 
@@ -24,8 +25,9 @@ class TemplateRenderTest extends TestCase
     protected function setUp(): void
     {
         $this->builderMock = $this->createMock(TemplateBuilderInterface::class);
+        $templateHelper = new TemplateHelper();
 
-        $this->render = new TemplateRender($this->builderMock);
+        $this->render = new TemplateRender($this->builderMock, $templateHelper);
     }
 
     #[WithoutErrorHandler]
@@ -68,6 +70,14 @@ class TemplateRenderTest extends TestCase
 
     private function createTemplateTransfer(): TemplateTransfer
     {
-        return TemplateHelper::getDefaultTemplateTransfer();
+
+        return new TemplateTransfer()->fromArray([
+            TemplateTransfer::CLASS_NAMESPACE => '\Default',
+            TemplateTransfer::CLASS_NAME => 'DefaultTransfer',
+            TemplateTransfer::IMPORTS => [
+                TransferEnum::ABSTRACT_CLASS->value,
+                TransferEnum::TRAIT->value,
+            ],
+        ]);
     }
 }
